requirements.txt
flask
openai
langchain-community
chromadb
python-dotenv
gunicorn
tiktoken
